Leitura dos dados em matriz

+++ src/heapsort/HeapSort.java
@@ -21,23 +21,23 @@ public class HeapSort {
 		//inicia-se  com o (tamanho do vetor -1) /2, /2 porque eh ordanado
 		//metade para cada lado da raiz e -1 porque um dos elementos eh a raiz
 		
-		for(long i=(vet.length-1)/2; i>=0; i--){
+		for(int i=(vet.length-1)/2; i>=0; i--){
 			heapTree(vet, i, vet.length);
 		}
 	}
 	
 	
-	public static void heapTree(long[] vet, long i, int tam) {
+	public static void heapTree(long[] vet, int i, int tam) {
 		
-		long j=2*i+1;
-		long k=j+1;
+		int j=2*i+1;
+		int k=j+1;
 		
 		//Compara os filhos e ordena, se filho 1(j) < filho 2(k) filho 1
 		//filho 2 passa a ser filho 1.
 		
 		if(j<tam && k<tam) {
 				
-			if(vet[(int)j]<vet[(int) k]) {
+			if(vet[j]<vet[k]) {
 				j=k;
 			}
 			
@@ -45,10 +45,10 @@ public class HeapSort {
 			//se for filho 1 vira raiz e raiz vira filho 1,
 			//ao final ocorre a reccusividade.
 			
-			if(vet[(int)j]>vet[(int)i]) {
-				long aux = vet[(int)j];
-				vet[(int)j]=vet[(int)i];
-				vet[(int)i]=aux;
+			if(vet[j]>vet[i]) {
+				long aux = vet[j];
+				vet[j]=vet[i];
+				vet[i]=aux;
 			
 					heapTree(vet, j, tam);
 			}
